добавлена PlayerInit

+++ sources/LevelOutput.c
@@ -2,9 +2,9 @@
 
 void LevelOutput(int **map)
 {
-	for(int row = 0; row < MAP_ROW_COUNT; row++)
+	for(size_t row = 0; row < MAP_ROW_COUNT; row++)
 	{
-		for(int col = 0; col < MAP_COL_COUNT; col++)
+		for(size_t col = 0; col < MAP_COL_COUNT; col++)
 		{
 			printf("%2d ", map[row][col]);
 		}
